the creation for the room with some validations

+++ src/main/java/com/hrp/hotel/microservice/hotel/infrastructure/outputadapters/db/JpaHotelRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JpaHotelRepository extends JpaRepository<HotelEntity, Integer> {
+public interface JpaHotelRepository extends JpaRepository<HotelEntity, Long> {
 
 }
